requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo[srv]==4.7.2
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.1
PyJWT==2.8.0
python-multipart==0.0.9
email-validator==2.1.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
itsdangerous==2.2.0
